split embedder
